[FE] REFACT: 클릭된 버튼 provider는 atom, 로그인 버튼 핸들러는 useOAuth로 이동 후 필요없는 props 삭제 #1760
